# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Postbase API",
    "version": "1.0.0",
    "description": "Schedule and manage social posts on X, LinkedIn, Bluesky, Mastodon and YouTube. Authenticate with an API key from Postbase → Developers, sent as `Authorization: Bearer pb_live_…`.",
    "contact": {
      "url": "https://www.postbase.so"
    },
    "license": {
      "name": "AGPL-3.0",
      "identifier": "AGPL-3.0-only"
    }
  },
  "servers": [
    {
      "url": "https://www.postbase.so/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/channels": {
      "get": {
        "summary": "List channels",
        "operationId": "listChannels",
        "tags": [
          "Channels"
        ],
        "responses": {
          "200": {
            "description": "Connected channels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "channels": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Channel"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/posts": {
      "get": {
        "summary": "List posts",
        "operationId": "listPosts",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "scheduled",
                "publishing",
                "published",
                "failed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Up to 100 posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PostSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a post",
        "description": "Save a draft, or schedule a post or thread. Send an Idempotency-Key so a retry can't post twice.",
        "operationId": "createPost",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "Repeat a request with the same key to get the first post back instead of a new one."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "body": {
                    "type": "string",
                    "description": "The post text, for a single post."
                  },
                  "thread": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Several posts to publish as a thread, in order. Takes precedence over body."
                  },
                  "channel_ids": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "description": "Channels to publish to, from GET /channels."
                  },
                  "channel_bodies": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "Text for particular channels instead of body, keyed by channel id. Each is a single post."
                  },
                  "scheduled_at": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "description": "When to publish: ISO 8601 with Z or an offset (2026-10-01T09:00:00+01:00). Omit or null to save a draft.",
                    "format": "date-time"
                  },
                  "schedule_in_minutes": {
                    "type": "number",
                    "description": "Publish this many minutes from now, on the server's clock. Instead of scheduled_at."
                  },
                  "media_ids": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "description": "Files from GET /media to attach, in order."
                  },
                  "media_urls": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    },
                    "description": "Public https links to images or a video, attached after media_ids. Each is downloaded into the media library (up to 10 per request, 200 MB each).",
                    "maxItems": 10
                  },
                  "youtube": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "privacy": {
                        "type": "string",
                        "enum": [
                          "public",
                          "unlisted",
                          "private"
                        ]
                      },
                      "made_for_kids": {
                        "type": "boolean",
                        "description": "Required to schedule to YouTube."
                      }
                    }
                  },
                  "idempotency_key": {
                    "type": "string",
                    "maxLength": 255,
                    "description": "Same as the Idempotency-Key header."
                  }
                }
              },
              "example": {
                "body": "Launching today 🚀",
                "channel_ids": [
                  "<channel-id>"
                ],
                "scheduled_at": "2026-10-01T09:00:00+01:00"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "post": {
                      "$ref": "#/components/schemas/CreatedPost"
                    }
                  }
                }
              }
            }
          },
          "200": {
            "description": "A repeat of an earlier request (same idempotency key)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "post": {
                      "$ref": "#/components/schemas/CreatedPost"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "No active plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/{id}/cancel": {
      "post": {
        "summary": "Cancel a post",
        "description": "Stop a scheduled post; it goes back to being a draft.",
        "operationId": "cancelPost",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string"
                    },
                    "still_sending": {
                      "type": "integer",
                      "description": "Channels already mid-send, which will finish"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the post's state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A pb_live_ API key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "A readable message"
          },
          "code": {
            "type": "string",
            "description": "A stable code, e.g. not_found"
          }
        },
        "required": [
          "error"
        ]
      },
      "Channel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "platform": {
            "type": "string",
            "enum": [
              "x",
              "linkedin",
              "bluesky",
              "mastodon",
              "youtube",
              "tiktok",
              "instagram",
              "facebook"
            ]
          },
          "handle": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "platform"
        ]
      },
...
      "PostSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "body": {
            "type": "string"
          },
          "scheduled_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "post_targets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "channel_id": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }
}